add stub icon and force macosx native compilation
Former-commit-id: 04b6f49daa0cd06c17b73f1c0a3dcd361dffc867 [formerly 456de06e6e82daef32215a55a97bfca4b5d7f085]
Former-commit-id: 35b3f769bef2632dc97bd862eb0caa90104d6849

+++ compile_webidebridge.sh
@@ -1,23 +1,28 @@
 # git submodule init
 # git submodule update
 
+if [ `uname` == Linux ]
+then
 cp -r arduino/tools_linux_64  arduino/tools
-goxc -os="linux" -arch="amd64" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
+CGO_ENABLED=1 goxc -os="linux" -arch="amd64" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
 rm -rf arduino/tools
 
 cp -r arduino/tools_linux_32  arduino/tools
-goxc -os="linux" -arch="386" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
+CGO_ENABLED=1 goxc -os="linux" -arch="386" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
 rm -rf arduino/tools
 
 cp -r arduino/tools_linux_arm  arduino/tools
-goxc -os="linux" -arch="arm" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
+CGO_ENABLED=1 goxc -os="linux" -arch="arm" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
 rm -rf arduino/tools
 
 cp -r arduino/tools_windows  arduino/tools
-goxc -os="windows" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=. -build-ldflags="-H=windowsgui"
+goxc -os="windows" -arch="386" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=. -build-ldflags="-H=windowsgui"
 rm -rf arduino/tools
+fi
 
+if [ `uname` == Darwin ]
+then
 cp -r arduino/tools_darwin  arduino/tools
-goxc -os="darwin" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
+CGO_ENABLED=1 goxc -os="darwin" -arch="amd64" --include="arduino/hardware,arduino/tools" -n="Arduino_WebIDE_Bridge" -d=.
 rm -rf arduino/tools
-
+fi
